refactor: rename HydraNode to Sybil
