+++ testing/helpers.go
@@ -6,16 +6,16 @@ import (
 	"github.com/ipfs/go-datastore"
 	"github.com/libp2p/go-libp2p-core/crypto"
 	"github.com/libp2p/go-libp2p-core/peer"
-	"github.com/libp2p/hydra-booster/node"
-	"github.com/libp2p/hydra-booster/node/opts"
+	"github.com/libp2p/hydra-booster/sybil"
+	"github.com/libp2p/hydra-booster/sybil/opts"
 )
 
 var defaults = []opts.Option{opts.Datastore(datastore.NewMapDatastore()), opts.BootstrapPeers(nil)}
 
 // SpawnNode a new Hydra nodes with an in memory datastore and 0 bootstrap peers by default.
 // It also waits for bootstrapping to complete.
-func SpawnNode(options ...opts.Option) (*node.HydraNode, error) {
-	nd, bsCh, err := node.NewHydraNode(append(defaults, options...)...)
+func SpawnNode(options ...opts.Option) (*sybil.Sybil, error) {
+	nd, bsCh, err := sybil.NewSybil(append(defaults, options...)...)
 	if err != nil {
 		return nil, err
 	}
@@ -34,8 +34,8 @@ func SpawnNode(options ...opts.Option) (*node.HydraNode, error) {
 }
 
 // SpawnNodes creates n new Hydra nodes with an in memory datastore and 0 bootstrap peers by default
-func SpawnNodes(n int, options ...opts.Option) ([]*node.HydraNode, error) {
-	var nodes []*node.HydraNode
+func SpawnNodes(n int, options ...opts.Option) ([]*sybil.Sybil, error) {
+	var nodes []*sybil.Sybil
 	for i := 0; i < n; i++ {
 		nd, err := SpawnNode(options...)
 		if err != nil {
